pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aegis-qa"
version = "0.1.0"
description = "Aegis — The AI Quality Control Plane"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Jack Blacketter"},
]

dependencies = [
    "typer>=0.12",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "rich>=13.7",
    "PyYAML>=6.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "httpx>=0.27",
    "aiosqlite>=0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=6.0",
    "ruff>=0.8",
    "mypy>=1.13",
    "types-PyYAML>=6.0",
    "pre-commit>=4.0",
]

[project.scripts]
aegis = "aegis_qa.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"aegis_qa" = [
    "landing/*.html",
    "landing/*.css",
    "landing/*.js",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--cov=aegis_qa --cov-fail-under=80"

[tool.ruff]
target-version = "py311"
line-length = 120
exclude = [
    ".venv",
    "venv",
    "build",
    "dist",
    ".eggs",
    "__pycache__",
]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
strict = false
disable_error_code = ["no-untyped-def"]

[tool.coverage.run]
source = ["aegis_qa"]

[tool.coverage.report]
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.",
    "if TYPE_CHECKING:",
]
